proact: attributes becomes properties for all vDOM nodes
The term `attributes` is more appropriate for elements, which represent HTML. But the term `properties` is more reflective of the intent: These are not just key-value pairs with values limited to text. No, these are live objects and functions, to be used by component at will, e.g., for sending and receiving events or for dispatching actions that modify the global store.