tests: protect code coverage by switching to relative module paths
esm 3.0.70 breaks code coverage if tests reach inside one of the
monorepo's packages. This is due to the fact that the same module
can be accessed through at least two file system paths, one being the
real path to the module and the other being the symbolic path through node_modules. While esm fixing this eventually seems likely, I don't want to
wait for the official fix since the underlying threat is so open.